refactor: use auth public exports
