+++ apps/frontend/src/app/main.ts
@@ -6,8 +6,8 @@ import { ToastService } from "primevue";
 import PrimeVue from "primevue/config";
 import Tooltip from "primevue/tooltip";
 import { createApp } from "vue";
-import { createAuthSession } from "@/features/auth/model/auth.session";
-import { useAuthStore } from "@/features/auth/model/auth.store";
+import { useAuthStore } from "@/features/auth";
+import { createAuthSession } from "@/features/auth/model";
 import { http } from "@/shared/api/http";
 import App from "./App.vue";
 import { primeVueOptions } from "./primevue-options";
