+++ apps/frontend/src/pages/login.vue
@@ -6,8 +6,7 @@ import { loginInputSchema } from "@recipes/shared";
 import { useToast } from "primevue";
 import { reactive } from "vue";
 import { useRoute, useRouter } from "vue-router";
-import { useLoginMutation } from "@/features/auth/api/auth.queries";
-import AuthPageShell from "@/features/auth/ui/AuthPageShell.vue";
+import { AuthPageShell, useLoginMutation } from "@/features/auth";
 
 definePage({
   meta: {
